+++ test/as you type.js
@@ -2,9 +2,15 @@ import chai, { expect } from 'chai'
 chai.should()
 
 import metadata from '../metadata.min'
-import as_you_typer, { close_dangling_braces, repeat } from '../source/as you type'
+import as_you_type_custom, { close_dangling_braces, repeat } from '../source/as you type'
 
-const as_you_type = as_you_typer(metadata)
+class as_you_type extends as_you_type_custom
+{
+	constructor(country_code)
+	{
+		super(country_code, metadata)
+	}
+}
 
 describe('as you type', () =>
 {